Major code reorganization
