+++ src/DictVectors/DictVectors.jl
@@ -9,20 +9,21 @@ concrete implementation of [`DVec`](@ref)
 """
 module DictVectors
 
-using Random, LinearAlgebra
+using Random
+using LinearAlgebra
+using StatsBase
+using ..StochasticStyles
+
+import ..StochasticStyles: deposit!
 import OrderedCollections: freeze
 import SplittablesBase
-export AbstractDVec, zero!, add!, deposit!, storage, walkernumber, localpart, freeze
+
+export AbstractDVec, zero!, add!, storage, walkernumber, freeze
 export DVec, InitiatorDVec
+
 
 export
     AbstractProjector, NormProjector, Norm2Projector, UniformProjector, Norm1ProjectorPPop
-export
-    compress!, NoCompression, ThresholdCompression, DoubleOrNothing, DoubleOrNothingWithTarget
-export
-    StochasticStyle, IsStochasticInteger, IsDeterministic, IsStochasticWithThreshold,
-    IsDynamicSemistochastic, IsExplosive, StyleUnknown,
-    IsDynamicSemistochasticPlus
 
 # The idea is to do linear algebra with data structures that are not
 # subtyped to AbstractVector, much in the spirit of KrylovKit.jl.
@@ -50,11 +51,8 @@ export
 # fill!(v, α)
 
 include("delegate.jl")
-include("compression.jl")
-include("stochasticstyle.jl")
 include("abstractdvec.jl")
 include("dvec.jl")
 include("initiators.jl")
-include("deprecated.jl")
 
 end # module
