+++ src/DictVectors/abstractdvec.jl
@@ -41,17 +41,6 @@ function Base.show(io::IO, dvec::AbstractDVec)
 end
 
 """
-    deposit!(w::AbstractDVec, add, val, parent::Pair)
-
-Add `val` into `w` at address `add`, taking into account initiator rules if applicable.
-`parent` contains the `address => value` pair from which the pair `add => val`
-was created. [`InitiatorDVec`](@ref) can intercept this and add its own functionality.
-"""
-function deposit!(w, add, val, _)
-    w[add] += convert(valtype(w), val)
-end
-
-"""
     storage(dvec) -> AbstractDict
 
 Return the raw storage associated with `dvec` as an `AbstractDict`. Used in MPI
@@ -332,13 +321,6 @@ walkernumber(w) = walkernumber(StochasticStyle(w), w)
 walkernumber(::StochasticStyle, w) = dot(Norm1ProjectorPPop(), w)
 # complex walkers as two populations
 # the following default is fast and generic enough to be good for real walkers and
-
-"""
-    localpart(dv) -> AbstractDVec
-
-Get the part of `dv` that is located on this MPI rank. Returns `dv` itself for `DictVector`s.
-"""
-localpart(dv) = dv # default for local data
 
 """
     FrozenDVec
