+++ src/Hamiltonians/offdiagonals.jl
@@ -50,27 +50,7 @@ julia> h = offdiagonals(H, addr)
  (BoseFS{6,3}((3, 3, 0)), -1.7320508075688772)
 ```
 """
-offdiagonals(h, a) = Offdiagonals(h, a)
-
-"""
-    random_offdiagonal(offdiagonals::AbstractOffdiagonals)
-    random_offdiagonal(ham::AbstractHamiltonian, add)
-
-Generate a single random excitation, i.e. choose from one of the accessible off-diagonal
-elements in the column corresponding to address `add` of the Hamiltonian matrix represented
-by `ham`. Alternatively, pass as argument an iterator over the accessible matrix elements.
-
-"""
-function random_offdiagonal(offdiagonals::AbstractOffdiagonals)
-    nl = length(offdiagonals) # check how many sites we could get_offdiagonal to
-    chosen = cRand(1:nl) # choose one of them
-    naddress, melem = offdiagonals[chosen]
-    return naddress, 1.0/nl, melem
-end
-
-function random_offdiagonal(ham::AbstractHamiltonian, add)
-    return random_offdiagonal(offdiagonals(ham, add))
-end
+offdiagonals(h::AbstractHamiltonian, a) = Offdiagonals(h, a)
 
 """
     Offdiagonals(h, address)
