+++ test/RMPI.jl
@@ -20,13 +20,13 @@ end
         dv = DVec(starting_address(ham)=>10; style=IsDynamicSemistochastic())
         v = MPIData(dv; setup)
         df, state = lomc!(ham,v)
-        @test size(df) == (100, 11)
+        @test size(df) == (100, 15)
     end
     # need to do mpi_one_sided separately
     dv = DVec(starting_address(ham)=>10; style=IsDynamicSemistochastic())
     v = RMPI.mpi_one_sided(dv; capacity = 1000)
     df, state = lomc!(ham,v)
-    @test size(df) == (100, 11)
+    @test size(df) == (100, 15)
 end
 
 @testset "sort_and_count!" begin
